pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rise-framework"
version = "2.0.0"
description = "A professional hierarchical multi-agent framework built on LangGraph"
readme = "README.md"
authors = [
    {name = "Rise Framework Team", email = "your.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["langchain", "langgraph", "multi-agent", "agents", "ai", "framework"]
dependencies = [
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.0.5",
    "langgraph>=0.0.20",
    "langchain-community>=0.0.20",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "typing-extensions>=4.0.0",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
mcp = [
    "langchain-mcp-adapters>=0.1.0",
]
rl = [
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/pinilDissanayaka/Rise-Framework-v2"
Documentation = "https://github.com/pinilDissanayaka/Rise-Framework-v2"
Repository = "https://github.com/pinilDissanayaka/Rise-Framework-v2"
"Bug Tracker" = "https://github.com/pinilDissanayaka/Rise-Framework-v2/issues"

[tool.setuptools]
packages = ["rise_framework"]

[tool.setuptools.package-data]
rise_framework = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=rise_framework --cov-report=html --cov-report=term"
